+++ RayTracer/Light/Light.cs
@@ -18,7 +18,7 @@ namespace RayTracer.Lighting
         public abstract Vector3 GetPointToLight(Point3 point);
 //        public abstract bool IsEffective(Point3 point, Geometry geometry, List<Geometry> geometries);
         public abstract bool IsEffective(Point3 point, Geometry geometry, Container bvh);
-        public abstract Double GetAttValue(Point3 point, Attenuation attenuation);
+        public abstract float GetAttValue(Point3 point, Attenuation attenuation);
 
     }
